Safeguard dividing by small values in PowerFilter. Temporary change in _rlnOpticGroup to get script to run (possible bug).

+++ src/aspire/utils/relion_interop.py
@@ -44,7 +44,7 @@ relion_metadata_fields = {
     "_rlnNrOfSignificantSamples": float,
     "_rlnNrOfFrames": int,
     "_rlnMaxValueProbDistribution": float,
-    "_rlnOpticsGroup": int,
+    "_rlnOpticsGroup": float,
     "_rlnOpticsGroupName": str,
 }
